Refactor to use common enums and util classes
