+++ crime-evidence/src/main/java/uk/gov/justice/laa/crime/evidence/staticdata/enums/EvidenceFeeRules.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.apache.commons.lang3.StringUtils;
+import uk.gov.justice.laa.crime.enums.EvidenceFeeLevel;
 import uk.gov.justice.laa.crime.evidence.dto.EvidenceFeeRulesDTO;
 
 import java.util.stream.Stream;
